[API Compatiblity] Support `paddle.exp`, `paddle.expm1`, `paddle.diagonal`, `paddle.round` (#76106)
* add api sink c++

* fix and add UT

* add round

* remove round

* fix
