+++ test/legacy_test/test_activation_op.py
@@ -6164,6 +6164,9 @@ class TestActivationAPI_Compatibility(unittest.TestCase):
     ACTIVATION_CONFIGS = [
         ("paddle.abs", np.abs, {'min_val': -1.0, 'max_val': 1.0}),
         ("paddle.log2", np.log2, {'min_val': 0.0, 'max_val': 8.0}),
+        ("paddle.exp", np.exp, {'min_val': -1.0, 'max_val': 1.0}),
+        ("paddle.expm1", np.expm1, {'min_val': -1.0, 'max_val': 1.0}),
+        ("paddle.round", np.round, {'min_val': -5.0, 'max_val': 5.0}),
     ]
 
     def setUp(self):
